feat: Create data module with LocalDoorDataSource interface (Phase 2.3)
First step of Phase 2.3 (Separate Data Modules). The data module
is pure Kotlin with no Android dependencies — it defines interfaces
for data sources that Repository implementations depend on.

- Created data/ module (kotlin("jvm"), depends on domain)
- Moved LocalDoorDataSource interface to com.chriscartland.garage.data
- androidApp's DatabaseLocalDoorDataSource implements the data interface
- All imports updated (DoorRepository, DoorRepositoryTest)

The data module is the abstraction layer between Repositories and
storage implementations (Room, Retrofit). Repositories depend on
data interfaces, never on Room/Retrofit directly.
