+++ AndroidGarage/androidApp/build.gradle.kts
@@ -225,6 +225,7 @@ room {
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":data"))
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
