+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/door/DoorRepository.kt
@@ -20,7 +20,7 @@ package com.chriscartland.garage.door
 import android.util.Log
 import com.chriscartland.garage.config.APP_CONFIG
 import com.chriscartland.garage.config.ServerConfigRepository
-import com.chriscartland.garage.db.LocalDoorDataSource
+import com.chriscartland.garage.data.LocalDoorDataSource
 import com.chriscartland.garage.domain.model.DoorEvent
 import com.chriscartland.garage.domain.model.DoorPosition
 import com.chriscartland.garage.domain.repository.DoorRepository
