+++ AndroidGarage/androidApp/src/test/java/com/chriscartland/garage/door/DoorRepositoryTest.kt
@@ -19,7 +19,7 @@ package com.chriscartland.garage.door
 
 import com.chriscartland.garage.config.ServerConfigRepository
 import com.chriscartland.garage.config.model.ServerConfig
-import com.chriscartland.garage.db.LocalDoorDataSource
+import com.chriscartland.garage.data.LocalDoorDataSource
 import com.chriscartland.garage.domain.model.DoorEvent
 import com.chriscartland.garage.domain.model.DoorPosition
 import com.chriscartland.garage.internet.CurrentEventDataResponse
